NB1: qcom-caf: gpu_tonemapper: improve checkEglError() MIME-Version: 1.0 Content-Type: text/plain; charset=UTF-8 Content-Transfer-Encoding: 8bit
Because eglGetError() returns only the error of the last called EGL
function (compared to glGetError() that may contain values for multiple
recorded errors), it is not necessary to do it in a loop.
Also the loop would only run for a seemingly random number of 5 times.

Also remove dead code & improve logging in checkGlError().

Test: make libgpu_tonemapper
Change-Id: I2c71bc92fa46c271b78d3171e089e2737542d770